+++ src/components/Accordion.vue
@@ -1,35 +1,34 @@
 <script lang="ts" setup>
 import {
-	Accordion,
-	AccordionContent,
-	AccordionItem,
-	AccordionTrigger,
+    Accordion,
+    AccordionContent,
+    AccordionItem,
+    AccordionTrigger,
 } from "@/components/ui/accordion";
-import type {HTMLAttributes} from "vue";
+import type { HTMLAttributes } from "vue";
 
 export interface Item {
-	value: string;
-	title: string;
-	content: string;
+    value: string;
+    title: string;
+    content: string;
 }
 
 export interface AccordionProps {
-	items: Item[];
-	type?: "single" | "multiple" | undefined;
-	collapsible?: boolean;
-	class?: HTMLAttributes["class"];
+    items: Item[];
+    type?: "single" | "multiple" | undefined;
+    collapsible?: boolean;
+    class?: HTMLAttributes[ "class" ];
 }
 
 const props = withDefaults(defineProps<AccordionProps>(), {
-	type: "single",
-	collapsible: true,
-	class: "w-full",
+    type: "single",
+    collapsible: true,
+    class: "w-full",
 });
 </script>
 
 <template>
-    <Accordion :type="props.type" :class="props.class" :collapsible="props.collapsible" data-testid="Accordion"
-        ignoreGlobalMixins>
+    <Accordion :type="props.type" :class="props.class" :collapsible="props.collapsible" data-testid="Accordion">
         <AccordionItem v-for="item in props.items" :key="item.value" :value="item.value">
             <AccordionTrigger>
                 <span v-translate>
