+++ src/config/directive/v-translate.ts
@@ -1,16 +1,16 @@
-import type {DirectiveBinding} from "vue";
-import {t} from "i18next";
+import type { DirectiveBinding } from "vue";
+import { t } from "i18next";
 
 function translateText(el: HTMLElement, binding: DirectiveBinding) {
-	const key = binding.value || el.innerText;
-	el.innerText = t(key);
+    const key = binding.value || el.innerText;
+    el.innerText = t(key);
 }
 
 export default {
-	mounted(el: HTMLElement, binding: DirectiveBinding) {
-		translateText(el, binding);
-	},
-	updated(el: HTMLElement, binding: DirectiveBinding) {
-		translateText(el, binding);
-	},
+    mounted(el: HTMLElement, binding: DirectiveBinding) {
+        translateText(el, binding);
+    },
+    updated(el: HTMLElement, binding: DirectiveBinding) {
+        translateText(el, binding);
+    },
 };
