+++ src/main.ts
@@ -1,7 +1,7 @@
 import "./assets/css/main.css";
 
-import {createApp} from "vue";
-import {createPinia} from "pinia";
+import { createApp } from "vue";
+import { createPinia } from "pinia";
 
 import App from "./App.vue";
 import router from "./router";
