+++ src/views/Home.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
-import {RouterLink} from "vue-router";
-import {useTranslation} from "@/config/composable/translateMixin";
+import { RouterLink } from "vue-router";
+import { useTranslation } from "@/config/composable/translate";
 
 useTranslation();
 </script>
